pyproject.toml
[tool.poetry]
name = "cvraft"
version = "0.1.0"
description = ""
authors = ["Manh-Ha VU <dev@manhhavu.com>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.0"
Markdown = "^3.3.6"
Jinja2 = "^3.0.3"
python-frontmatter = "^1.0.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"
pytest = "^7.0.1"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cvraft = "cvraft.cli:app"
